refactor(api): rename evaluation/assessments route file
