+++ api/tests/evaluation/integration/application/assessments/assessment-controller_test.js
@@ -1,6 +1,6 @@
 import sinon from 'sinon';
 
-import { assessmentsRoute as moduleUnderTest } from '../../../../../src/evaluation/application/assessments/index.js.js';
+import { assessmentsRoute as moduleUnderTest } from '../../../../../src/evaluation/application/assessments/index.js';
 import { assessmentAuthorization } from '../../../../../src/evaluation/application/pre-handlers/assessment-authorization.js';
 import { evaluationUsecases } from '../../../../../src/evaluation/domain/usecases/index.js';
 import { expect } from '../../../../test-helper.js';
